update : slicer with thunk
